rewrite Generalize as class
